Fix level requirement for skill gems

+++ PyPoE/poe/file/specification/data/stable.py
@@ -9574,7 +9574,7 @@ specification = Specification({
             ),
             Field(
                 name='PlayerLevelReq',
-                type='int',
+                type='float',
             ),
             Field(
                 name='CostMultiplier',
